Continued work on V0
Still a ways away from launch. At some point I need to figure out the story, make it so that a player cannot simply skip over my code. More immidately, I need to figure out combat, how to add the sword to inventory, and how to make the sword change the stats of the player.